MET-60: Rename folders
