+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/config/routing.yml
@@ -1,2 +1,2 @@
 akeneo_measurements_settings_index:
-    path: '/configuration/measurements'
+    path: '/configuration/measurement'
