+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/bridge/controller/settings.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import {dependencies} from '../dependencies';
 import ReactController from '../react/react-controller';
-import App from 'akeneomeasure/js/app/index';
+import App from 'akeneomeasure/index';
 
 const mediator = require('oro/mediator');
 
 class SettingsController extends ReactController {
   reactElementToMount() {
-    return <App dependencies={dependencies}/>;
+    return <App dependencies={dependencies} />;
   }
 
   routeGuardToUnmount() {
