+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/bridge/dependencies.ts
