+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/bridge/react/react-controller.ts
