+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/bridge/react/react-element-helper.ts
