+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/shared/components/PageContent.tsx
