+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/shared/legacy/legacy-context.ts
@@ -1,8 +1,8 @@
 import {createContext} from 'react';
-import {ViewBuilder} from 'akeneomeasure/js/app/shared/legacy/pim-view/view-builder';
+import {ViewBuilder} from 'akeneomeasure/shared/legacy/pim-view/view-builder';
 
 export type LegacyContextValue = {
-    viewBuilder?: ViewBuilder;
-}
+  viewBuilder?: ViewBuilder;
+};
 
 export const LegacyContext = createContext<LegacyContextValue>({});
