+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/shared/legacy/pim-view/view-builder.ts
