+++ src/Akeneo/Tool/Bundle/MeasureBundle/Resources/public/shared/theme.ts
